Add fileParseController and multer for server
Server.js uses multer to handle file uploads. Multer extracts the data
from the request, puts it into an obj based on field names, and stores
files TEMPORARILY in memory with multer.memoryStorage.

The fileParseController, in turn, concurrently parses the files using
Promise.all. So far, text files are functioning (their buffer is turned
into a string). The PDF processing is not working as of this commit.