Add compareController & rename fileParse -> parse
The compareController takes in the parsed files and examines them line
by line. Presently, the result of the comparison is an object:
{
  isEqual: boolean,
  differences: Array<{line: int, file1: str, file2: str}>
}

I think this is helpful, but it might be too verbose.

I also renamed fileParseController to parseController to be more in line
with the naming convention of compareController